preparing to test modes

+++ roles/carousel/lighting.py
@@ -81,7 +81,7 @@ def set_group(id, level):
 def set_spoke(id, level):
     radial_spoke = radial_spokes[id]
     for led in radial_spoke:
-        pins[led].duty_cycle = level*65536
+        pins[led].duty_cycle = int(level*65536)
 
 def fade_spoke(id, start_level, end_level, fade_period = 0.1):
     animation_frame_period = 0.01
